Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -1,13 +1,13 @@
 #include <stdio.h>
 
 int main() {
-    int a, b, c, avg;
+    float a, b, c, avg;
     printf("");
-    scanf("%d", &a);
-    scanf("%d", &b);
-    scanf("%d", &c);
+    scanf("%f", &a);
+    scanf("%f", &b);
+    scanf("%f", &c);
 
     avg= (a + b + c)/3;
-    printf("Average: %d", avg);
+    printf("Average: %.2f", avg);
     return 0;
 }
